use redux for rule form state
